Mudança de id

+++ src/test/kotlin/com/github/thaynarasilvapinto/simuladorbanco/ContaControllerTest.kt
@@ -58,8 +58,10 @@ class ContaControllerTest {
     fun tearDown() {
         clienteService.delete(joao.id)
         val extrato = operacaoService.findAllContaOrigem(joaoConta)
-        for (i in extrato.indices) {
-            operacaoService.delete(extrato[i].idOperacao)
+        if(extrato != null) {
+            for (i in extrato!!.indices) {
+                operacaoService.delete(extrato[i].idOperacao)
+            }
         }
         contaService.delete(joaoConta.id)
     }
